refactor: AI 모델 조회 구조 및 정보 필드 UI 정리
- 기존 aiModelQueries 파일 제거
- 모델 정보와 성능 평가 query option을 기능별 파일로 분리
- 성능 평가 mutation 로직을 별도 mutation 파일로 분리
- 모델 정보 섹션의 렌더링 분기 구조 정리
- ModelInfoSection의 grid UI를 InfoFieldGrid로 분리
